Fix vertical squashing of the panel icons
The teapot and especially the progress circle icon in the panel are
squashed with GNOME-3.32. The problem appears to be that the padding
which we are requesting gets applied not only to the surface-size
which we are being passed (which is 20 px wide for our 24 px size),
but gets applied a second time using scaling squashing the image
and making the progress circle non round.

This commit fixes this by removing the padding, as gnome shell
automatically adds plenty of padding and directly requesting 20x20 as
surface size, which results in a 20x27 surface as we are given the full
height of the panel as height, but our own padding code takes care of
that.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>