Fix TeaTime not working with gnome-shell 3.35.90 and newer
This commit contains the minimum amount of changes for TeaTime to mostly
work with gnome-shell 3.35.90 and newer.

There is one big remaining issue after this, the fullscreen notifications
are not working anymore. Fixing this unfortunately is non trivial,
so I plan to remove it in a separate commit.

There are also some JS warnings, I will fix those in separate commits.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>