Fix indendation
After the "Fix TeaTime not working with gnome-shell 3.35.90 and newer"
commit, a bunch of blocks no longer have correct indentation.

These blocks were left as is on purpose to make the diff of that commit
easier to read.

This commit fixes the indentation of these blocks. This commit makes no
changes other then the indentation changes and as such contains no
functional changes.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>